console.lua: improve the hovered item calculation with background-box
Follow up to c438732b23 which improved the calculation with the default
outline-and-shadow. We can make the calculation accurate for
background-box too without making semitransparent rectangle backgrounds
overlap by adding margin between items.

We could calculate the height of each item to make them perfectly
adjacent by rendering each one with compute_bounds, but that takes 15ms
per render, so 20 lines would take 300ms, which is too slow.

Instead add a fixed 0.1 * opts.font_size to each item's height. This
ensures the backgrounds don't overlap with tall CJK text or track
circles and --osd-shadow-offset=0. Add this to outline-and-shadow too
since it looks better with some margin between the items. Then add the
rectangle offset to the height depending on the border style.